omdb_api/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
pandas==2.2.0
requests==2.31.0
urllib3==1.26.6